refactor: consolidate workshop components and extract shared utilities
