+++ src/components/workshop/cursor/Conclusion.tsx
@@ -1,6 +1,6 @@
 'use client'
 import {motion} from 'framer-motion'
-import {fadeInUp} from './animations'
+import {fadeInUp} from '../shared/animations'
 
 export default function Conclusion() {
   return (
