--- src/components/workshop/cursor/styles.css
@@ -1,18 +1,0 @@
-.pattern-dots {
-  @apply bg-white dark:bg-gray-900;
-  background-image: linear-gradient(
-      to bottom,
-      rgba(255, 255, 255, 0.9),
-      transparent
-    ),
-    radial-gradient(rgba(0, 0, 0, 0.1) 2px, transparent 2px);
-  background-size: 100% 100%, 30px 30px;
-  background-position: 0 0, 15px 15px;
-}
-
-.dark .pattern-dots {
-  background-image: linear-gradient(to bottom, rgba(0, 0, 0, 0.15), transparent),
-    radial-gradient(rgba(255, 255, 255, 0.15) 2px, transparent 2px);
-  background-size: 100% 100%, 30px 30px;
-  background-position: 0 0, 15px 15px;
-}
