+++ src/components/workshop/cursor/team/hero.tsx
@@ -1,7 +1,7 @@
 'use client'
 import Link from 'next/link'
 import {motion} from 'framer-motion'
-import {fadeInUp, scaleIn} from '../animations'
+import {fadeInUp, scaleIn} from '../../shared/animations'
 import {useState, useEffect} from 'react'
 import '../styles.css'
 import Image from 'next/image'
