+++ src/pages/workshop/cursor/invoice/[merchantChargeId].tsx
@@ -5,8 +5,8 @@ import {format, fromUnixTime} from 'date-fns'
 import {ChevronLeft, MailIcon, Receipt} from 'lucide-react'
 import Stripe from 'stripe'
 import Image from 'next/image'
-import {InvoiceCustomText} from '@/components/workshop/cursor/invoices/invoice-custom-text'
-import {InvoicePrintButton} from '@/components/workshop/cursor/invoices/invoice-print-button'
+import {InvoiceCustomText} from '@/components/workshop/shared/invoices/invoice-custom-text'
+import {InvoicePrintButton} from '@/components/workshop/shared/invoices/invoice-print-button'
 import {Button, Input} from '@/ui'
 import {redirect} from 'next/navigation'
